Creating master IMU function
General clean up for the IMU files and creating Master function
