+++ IMUFunctions.py
@@ -1,0 +1,326 @@
+# -*- coding: utf-8 -*-
+"""
+Shared functions for IMU-based biomechanical analyses. These functions
+are used across multiple IMU processing scripts.
+
+@author: Eric.Honert
+"""
+
+import numpy as np
+from numpy import cos, sin, arctan2
+import scipy
+import scipy.interpolate
+import scipy.signal as sig
+from scipy.integrate import cumulative_trapezoid as cumtrapz
+import matplotlib.pyplot as plt
+
+
+def filtIMUsig(sig_in, cut, t):
+    """
+    Filter 3 axes of an IMU signal (X,Y,Z) with a 2nd order butterworth filter
+    at the specified cut-off frequency
+
+    Parameters
+    ----------
+    sig_in : numpy array (Nx3)
+        acceleration or gyroscope signal
+    cut : float
+        cut-off frequency
+    t : numpy array
+        time (sec)
+
+    Returns
+    -------
+    sig_out : numpy array (Nx3)
+        filtered signal at the specified cut-off frequency
+
+    """
+    # Set up a 2nd order low pass buttworth filter
+    freq = 1/np.mean(np.diff(t))
+    w = cut / (freq / 2) # Normalize the frequency
+    b, a = sig.butter(2, w, 'low')
+    # Filter the IMU signals
+    sig_out = np.array([sig.filtfilt(b, a, sig_in[:,jj]) for jj in range(3)]).T
+    return(sig_out)
+
+def delimitTrialIMU(SegSig):
+    """
+    Function to crop the data using interactive ginput selection
+
+    Parameters
+    ----------
+    SegSig : numpy array or list
+        The signal that will be plotted to allow for segmentation
+
+    Returns
+    -------
+    segidx : array
+        Segmentation indices (start and end)
+
+    """
+    print('Select 2 points: the start and end of the trial')
+    fig, ax = plt.subplots()
+    ax.plot(SegSig, label = 'Segmenting Signal')
+    fig.legend()
+    pts = np.asarray(plt.ginput(2, timeout=40))
+    plt.close()
+    segidx = pts[:,0]
+    return(segidx)
+
+def align_fuse_extract_IMU(LGdat, HGdat):
+    """
+    Function to align, fuse, and extract all IMU information
+
+    The high-g accelerometer is interpolated to the low-g accelerometer using
+    the UNIX time-stamps in the innate files. The two accelerometers are then
+    aligned using a cross correlation. Typically this results in a 1 frame
+    phase shift
+
+    Parameters
+    ----------
+    LGdat : dataframe
+        low-g data frame that is extracted as raw data from Capture.U. This
+        dataframe contains both the low-g accelerometer and the gyroscope.
+    HGdat : dataframe
+        high-g data frame that is extracted as raw data from Capture.U.
+
+    Returns
+    -------
+    LGtime : numpy array (Nx1)
+        time (UNIX) from the low-g file
+    acc : numpy array (Nx3)
+        X,Y,Z fused (between low-g and high-g) acceleration from the IMU
+    gyr : numpy array (Nx3)
+        X,Y,Z gyroscope from the IMU
+
+    """
+    # LGdat: low-g dataframe (contains accelerometer and gyroscope)
+    # HGdat: high-g data frame (contains only accelerometer)
+
+    # Need to align the data
+    # 1: get the data collection frequency from the low-g acclerometer
+    acc_lg = np.array(LGdat.iloc[:,2:5])
+    gyr = np.array(LGdat.iloc[:,5:])
+
+    dum = np.array(HGdat.iloc[:,2:])
+
+    HGtime = np.array(HGdat.iloc[:,0])
+    LGtime = np.array(LGdat.iloc[:,0])
+
+    idx = (LGtime < np.max(HGtime))*(LGtime >= np.min(HGtime))
+    LGtime = LGtime[idx]
+    acc_lg = acc_lg[idx,:]
+    gyr = gyr[idx,:]
+
+    # Create an empty array to fill with the resampled/downsampled high-g acceleration
+    resamp_HG = np.zeros([len(LGtime),3])
+
+    for jj in range(3):
+        f = scipy.interpolate.interp1d(HGtime,dum[:,jj])
+        resamp_HG[:,jj] = f(LGtime)
+
+    # Cross-correlate the y-components
+    corr_arr = sig.correlate(acc_lg[:,1],resamp_HG[:,1],mode='full')
+    lags = sig.correlation_lags(len(acc_lg[:,1]),len(resamp_HG[:,1]),mode='full')
+    lag = lags[np.argmax(corr_arr)]
+
+    if lag > 0:
+        LGtime = LGtime[lag:]
+        gyr = gyr[lag:,:]
+        acc_lg = acc_lg[lag:,:]
+        resamp_HG = resamp_HG[:-lag,:]
+
+    elif lag < 0:
+        LGtime = LGtime[:lag]
+        gyr = gyr[:lag,:]
+        acc_lg = acc_lg[:lag,:]
+        resamp_HG = resamp_HG[-lag:,:]
+
+    acc = acc_lg
+
+    # Find when the data is above/below 16G and replace with high-g accelerometer
+    for jj in range(3):
+        idx = np.abs(acc[:,jj]) > (9.81*16-0.1)
+        acc[idx,jj] = resamp_HG[idx,jj]
+
+    return [LGtime,acc,gyr]
+
+def findRotToLab(accel_vec):
+    """
+    Function to find the rotation of the foot flat acceleration vector to the
+    defined lab gravity coordinate system.
+
+    Parameters
+    ----------
+    accel_vec : numpy array
+        3x1 vector of the x,y,z acceleration at foot flat
+
+    Returns
+    -------
+    theta : numpy array
+        3x1 vector for the rotation from the foot flat acceleration to the lab
+        coordinate system (deg)
+
+    """
+    iGvec = accel_vec
+    iGvec = iGvec/np.linalg.norm(iGvec)
+    # Define the lab coordinate system gravity vector
+    lab_Gvec = np.array([1,0,0]).T
+    #______________________________________________________________________
+    # Compute the rotation matrix
+    C = np.cross(lab_Gvec,iGvec)
+    D = np.dot(lab_Gvec,iGvec)
+    Z = np.array([[0,-C[2],C[1]],[C[2],0,-C[0]],[-C[1],C[0],0]])
+    # Note: test the rotation matrix (R) that the norm of the colums and the rows is 1
+    R = np.eye(3)+Z+(Z@Z)*(1-D)/np.linalg.norm(C)**2
+    # Compute the rotation angles
+    theta = np.zeros(3) # Preallocate
+    theta[1] = arctan2(-R[2,0],np.sqrt(R[0,0]**2+R[1,0]**2))
+    # Conditional statement in case theta[1] is +/- 90 deg (pi/2 rad)
+    if theta[1] == np.pi/2:
+        theta[2] = 0
+        theta[0] = arctan2(R[0,1],R[1,1])
+    elif theta[1] == -np.pi/2:
+        theta[2] = 0
+        theta[0] = -arctan2(R[0,1],R[1,1])
+    else:
+        theta[2] = arctan2(R[1,0]/cos(theta[1]),R[0,0]/cos(theta[1]))
+        theta[0] = arctan2(R[2,1]/cos(theta[1]),R[2,2]/cos(theta[1]))
+    theta = theta*180/np.pi
+    return theta
+
+def intp_strides(var, landings, GS):
+    """
+    Function to interpolate the variable of interest across a stride
+    (from foot contact to subsequent foot contact) in order to plot the
+    variable of interest over top each other
+
+    Parameters
+    ----------
+    var : list or numpy array
+        Variable of interest. Can be taken from a dataframe or from a numpy array
+    landings : list
+        Foot contact indices
+    GS : list or numpy array
+        Indices of good strides to interpolate
+
+    Returns
+    -------
+    intp_var : numpy array
+        Interpolated variable to 101 points with the number of columns dictated
+        by the number of strides.
+
+    """
+    # Preallocate
+    intp_var = np.zeros((101,len(GS)-1))
+    # Index through the strides
+    for ii in range(len(GS)-1):
+        dum = var[landings[GS[ii]]:landings[GS[ii]+1]]
+        f = scipy.interpolate.interp1d(np.arange(0,len(dum)),dum)
+        intp_var[:,ii] = f(np.linspace(0,len(dum)-1,101))
+
+    return intp_var
+
+def computeRunSpeedIMU(acc, gyr, HS, MS, GS, t):
+    """
+    Estimate running speed from IMU signals.
+
+    This function rotates the acceleration signal into an inertial coordinate
+    system based on the integrand of the gyroscope signal. This rotation is
+    performed in order to account for the gravitational acceleration. The
+    corrected inertial coordinate system acceleration is then double
+    integrated to estimate the displacement. The integrations are linearly
+    corrected. Note that all time continuous signals span from foot contact
+    to the subsequent midstance. The final displacement calculation is from
+    foot contact to the subsequent foot contact.
+
+    Parameters
+    ----------
+    acc : numpy array (Nx3)
+        X,Y,Z acceleration from the IMU
+    gyr : numpy array (Nx3)
+        X,Y,Z gyroscope from the IMU
+    HS : list
+        Heel-strike (foot contact events) indices
+    MS : list
+        Mid-stance indices
+    GS : list
+        Indices that indicate good strides
+    t : numpy array (Nx1)
+        time (seconds)
+
+    Returns
+    -------
+    run_speed : list
+        Step-by-step running speed
+
+    """
+    # Set up a 2nd order 50 Hz low pass buttworth filter
+    freq = 1/np.mean(np.diff(t))
+    w = 50 / (freq / 2) # Normalize the frequency
+    b, a = sig.butter(2, w, 'low')
+
+    # Filter the IMU signals
+    acc_filt = np.array([sig.filtfilt(b, a, acc[:,jj]) for jj in range(3)]).T
+    gyr_filt = np.array([sig.filtfilt(b, a, gyr[:,jj]) for jj in range(3)]).T
+
+    # Indicate the number of frames to examine the gravity during midstance
+    MS_frames = 20
+
+    run_speed = []
+
+    for count, jj in enumerate(GS):
+        # Obtain the acceleration and gyroscope from foot contact (HS) to the
+        # subsequent midstance (MS). MS is needed for linear drift correction
+        acc_stride_plus = acc_filt[HS[jj]:MS[jj+1],:]
+        gyr_stride_plus = gyr_filt[HS[jj]:MS[jj+1],:]
+        time_stride_plus = t[HS[jj]:MS[jj+1]]
+        time_stride = t[HS[jj]:HS[jj+1]]
+
+        # Need the midstance and next foot contact indices
+        MS_idx = MS[jj]-HS[jj]
+        HS_idx = HS[jj+1]-HS[jj]
+
+        # Compute the unit vector pointing in the approximate direction of gravity
+        Fflat_accel = np.mean(acc_stride_plus[MS_idx:MS_idx+MS_frames,:],axis = 0)/np.mean(np.linalg.norm(acc_stride_plus[MS_idx:MS_idx+MS_frames,:],axis = 1)).T
+
+        # Rotation to the gravity vector: provides initial condition for gyro integration
+        thetai = findRotToLab(Fflat_accel)
+
+        # Integrate and linearly correct the gyroscope
+        theta = cumtrapz(gyr_stride_plus,time_stride_plus,initial=0,axis=0)
+        theta = theta-theta[MS_idx,:] # Midstance should be zero
+        # Linear correction
+        slope = theta[-1,:]/(len(theta)-MS_idx-1)
+        drift = (slope*np.ones([len(theta),3]))*(np.array([range(len(theta)),range(len(theta)),range(len(theta))])).T
+        drift = drift-drift[MS_idx,:]
+        theta = theta - drift
+        # Gyro integration initial condition
+        if count == 0:
+           theta = theta-thetai
+           thetai_up = thetai
+        else:
+           theta = theta-(thetai+thetai_up)/2 # average between the previous step
+           thetai_up = thetai
+
+        # Convert to radians
+        theta = theta*np.pi/180
+        # Rotate the accelerometer into the lab coordinate system (LCS)
+
+        # Note: this rotation has been converted to list comprehension for speed purposes
+        acc_stride_plus_LCS = [(np.array([[cos(ang[2]),-sin(ang[2]),0],[sin(ang[2]),cos(ang[2]),0],[0,0,1]])@np.array([[1,0,0],[0,cos(ang[0]),-sin(ang[0])],[0,sin(ang[0]),cos(ang[0])]])@np.array([[cos(ang[1]),0,sin(ang[1])],[0,1,0],[-np.sin(ang[1]),0,cos(ang[1])]])@acc_stride_plus[kk,:].T).T-np.array([9.81,0,0]) for kk,ang in enumerate(theta)]
+
+        # Integrate to get IMU velocity in the LCS
+        IMU_vel = cumtrapz(acc_stride_plus_LCS,time_stride_plus,initial=0,axis=0)
+        # Initial condition: IMU velocity is zero at midstance
+        IMU_vel = IMU_vel-IMU_vel[MS_idx,:]
+        # Remove drift
+        slope = IMU_vel[-1,:]/(len(IMU_vel)-MS_idx-1)
+        drift = (slope*np.ones([len(IMU_vel),3]))*(np.array([range(len(IMU_vel)),range(len(theta)),range(len(IMU_vel))])).T
+        drift = drift-drift[MS_idx,:]
+        IMU_vel = IMU_vel-drift
+        # Integrate to get step length in the LCS
+        IMU_SL = np.trapezoid(IMU_vel[0:HS_idx,:],time_stride,axis=0)
+        run_speed.append(np.linalg.norm(IMU_SL)/(time_stride[-1]-time_stride[0]))
+
+    return run_speed
